add index.ts file for icons, change icons structure a bit
