+++ src/js/interfaces/SVGprops.ts
@@ -1,0 +1,6 @@
+export interface SVGprops {
+  size: number;
+  color?: string;
+  fill?: string;
+  direction?: string;
+}
