+++ src/js/interfaces/index.ts
@@ -2,3 +2,4 @@ export * from "./List";
 export * from "./Person";
 export * from "./Item";
 export * from "./Store";
+export * from "./SVGprops";
