+++ src/js/view/icons/ArrowIcon/ArrowIcon.css
@@ -1,0 +1,15 @@
+.arrow-icon_up {
+    transform: rotate(0deg);
+}
+
+.arrow-icon_down {
+    transform: rotate(180deg);
+}
+
+.arrow-icon_right {
+    transform: rotate(90deg);
+}
+
+.arrow-icon_left {
+    transform: rotate(-90deg);
+}
