+++ src/js/view/icons/CalendarIcon.tsx
@@ -1,14 +1,8 @@
 import React, { FunctionComponent } from "react";
+import { SVGprops } from "../../interfaces";
 
-export interface SVGprops {
-  size: number;
-  color?: string;
-  fill?: string;
-}
-
-// eslint-disable-next-line react/function-component-definition
 export const CalendarIcon: FunctionComponent<SVGprops> = (props) => {
-  const { size, color, fill } = props;
+  const { size, color = "white", fill = "none" } = props;
   return (
     <svg
       width={size.toString()}
@@ -61,9 +55,4 @@ export const CalendarIcon: FunctionComponent<SVGprops> = (props) => {
       />
     </svg>
   );
-};
-
-CalendarIcon.defaultProps = {
-  color: "white",
-  fill: "none",
 };
