+++ src/js/view/icons/ShareIcon.tsx
@@ -1,14 +1,8 @@
 import React, { FunctionComponent } from "react";
+import { SVGprops } from "../../interfaces";
 
-export interface SVGprops {
-  size: number;
-  color?: string;
-  fill?: string;
-}
-
-// eslint-disable-next-line react/function-component-definition
 export const ShareIcon: FunctionComponent<SVGprops> = (props) => {
-  const { size, color, fill } = props;
+  const { size, color = "white", fill = "none" } = props;
   return (
     <svg
       width={size.toString()}
@@ -55,9 +49,4 @@ export const ShareIcon: FunctionComponent<SVGprops> = (props) => {
       />
     </svg>
   );
-};
-
-ShareIcon.defaultProps = {
-  color: "white",
-  fill: "none",
 };
