+++ src/js/view/icons/ZoomIcon.tsx
@@ -1,14 +1,8 @@
 import React, { FunctionComponent } from "react";
+import { SVGprops } from "../../interfaces";
 
-export interface SVGprops {
-  size: number;
-  color?: string;
-  fill?: string;
-}
-
-// eslint-disable-next-line react/function-component-definition
 export const ZoomIcon: FunctionComponent<SVGprops> = (props) => {
-  const { size, color, fill } = props;
+  const { size, color = "white", fill = "none" } = props;
   return (
     <svg
       width={size.toString()}
@@ -31,9 +25,4 @@ export const ZoomIcon: FunctionComponent<SVGprops> = (props) => {
       />
     </svg>
   );
-};
-
-ZoomIcon.defaultProps = {
-  color: "white",
-  fill: "none",
 };
