+++ src/js/view/icons/index.ts
@@ -1,0 +1,10 @@
+export * from "./ArrowIcon/ArrowIcon";
+export * from "./CalendarIcon";
+export * from "./CloseIcon";
+export * from "./EditIcon";
+export * from "./DeleteIcon";
+export * from "./MinusIcon";
+export * from "./PlusIcon";
+export * from "./KebabIcon";
+export * from "./ShareIcon";
+export * from "./ZoomIcon";
